Extract frontend-neutral provider orchestration into agent-runtime
