--- packages/agent-cli/src/Agent/CLI/ProviderRuntime.hs
@@ -1,17 +1,0 @@
-module Agent.CLI.ProviderRuntime
-    ( withProviderRuntime
-    , ProviderConfig(..)
-    , OpenRouterConfig(..)
-    , ProviderHost(..)
-    , ProviderCompaction(..)
-    , ProviderRuntime(..)
-    ) where
-
-import Agent.CLI.Runtime.Orchestration.Providers (withProviderRuntime)
-import Agent.CLI.Runtime.Orchestration.Providers.Types (
-    OpenRouterConfig(..),
-    ProviderCompaction(..),
-    ProviderConfig(..),
-    ProviderHost(..),
-    ProviderRuntime(..),
- )
