+++ packages/agent-cli/test/Agent/CLI/TurnSpec.hs
@@ -9,7 +9,7 @@ import Agent.CLI.Turn
     )
 import Agent.CLI.TurnState
 import Agent.CLI.NativeRuntime (StartupFailure(..), exitFailedTurn)
-import Agent.CLI.Compaction (AutomaticCompactionBoundary(..))
+import Agent.Runtime.Compaction (AutomaticCompactionBoundary(..))
 import Agent.Error (ApiError(..))
 import Agent.Loop
     ( ImageAttachment(..)
