+++ packages/agent-runtime/package.nix
@@ -27,11 +27,11 @@ mkDerivation {
     wai warp
   ];
   testHaskellDepends = [
-    aeson agent-accounts agent-core agent-json agent-openai
-    agent-responses agent-responses-types agent-store agent-tools async
-    base bytestring containers directory filepath hspec http-client
-    http-types network QuickCheck safe-exceptions stm text time unix
-    wai warp
+    aeson agent-accounts agent-connectivity agent-core agent-json
+    agent-openai agent-openrouter agent-responses agent-responses-types
+    agent-store agent-tools agent-xai async base bytestring containers
+    directory filepath hspec http-client http-types network QuickCheck
+    safe-exceptions stm text time transformers unix wai warp
   ];
   description = "Frontend-independent conversation and turn lifecycle";
   license = lib.meta.getLicenseFromSpdxId "MIT";
